Added INDEPTH specifications as data
I have added the list of event codes as well as a data frame of verbal
autopsy results as data objects, so they are visible outside of the
validation code. I ahve also added documentation for both objects in a
single dummy R file, called dataDocumentation.R, which just has roxygen2
documentation and a couple of NULLs.