Introduced pipe-based variable testing
Changed from the previous incarnation, which had a mixture of variable-
and record-tests pre/post-pre-processing to introduce pure variable
testing as first step (pre-preprocessing). The idea is that we only report
on one variable at a time, w/o references between variables (as far as
possible). The component tests are now designed for piping, though we
introduce the pipe opertor via dplyr instead of magrittr directly.